FIX SSL ERROR
+add class filter_link
+error handler added new + now writes link index (more info for testing)

There was a problem with visiting sites where there was an unconfirmed SSL certificate (due to sanctions on Russia, most Russian sites could not renew free self-written certificates). To fix it, it was decided to temporarily add a filter that, if the SSL fails, will change HTTPS links to HTTP and change port 443 to 80.