fix crash program
+ added timer (in find_mail)
+ Algorithm for accessing the site has changed.

Errors were thrown that the handler could not catch. I had to use an adapter with Retry. It was adapted for fast loading and error interception (it doesn’t always work, but if it doesn’t work, it reads the site as it is in the end and if it didn’t find anything, it gives an error about leaving the array).